add delete TestCase button
